% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustDynamicProxy.R
\name{adjustDynamicProxy}
\alias{adjustDynamicProxy}
\title{Adjust the map.tib between iid and iid_proxy}
\usage{
adjustDynamicProxy(map.tib, asOfDate, li)
}
\arguments{
\item{map.tib}{the map between investment iid and proxy iid.}

\item{asOfDate}{the reference date used to adjust the map.tib.}

\item{li}{the exported list of investments containing the PXX tibbles.}
}
\value{
a tibble identical to the input map.tib where the entries of
   the field iid_proxy equal to 'by_rat_mat' have been replaced by the 
   correct proxy according to the mapByCcy.l component of the rating.l
   created by the createRatingData function. For the modified rows the
   field type_proxy is set equal to 'idx'.
}
\description{
Adjust the map.tib between iid and iid_proxy by replacing the 'by_rat_mat'
       values of the iid_proxy with the associated iid_proxy (see the map
       produced by the createRatingData function).
}
