% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_OC_SRRI.R
\name{computeOC_SRRI}
\alias{computeOC_SRRI}
\title{Compute the SRRI from a timeseries of returns and an asOfDate}
\usage{
computeOC_SRRI(iid, retXts.l, asOfDate, show_info = TRUE)
}
\arguments{
\item{iid}{the INTERNAL_ID of the investment object.}

\item{retXts.l}{a list of xts weekly returns, where each component is the result of createWeeklyReturnsTs.}

\item{asOfDate}{the reference date.}

\item{show_info}{logical. If TRUE the 'INFO: ...' messages are printed to the console.}
}
\value{
a list with the following fields:

		- ret.xts the xts of weekly returns
		- dataRange the range of ret.xts
		- nObs the number of weekly returns 
		- realizedYearlyVol the realized yearly volatility
		- srri the synthetic risk and reward indicator
		- info a named vector of character strings
}
\description{
Compute the SRRI from a timeseries of returns and an asOfDate
}
