% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_isin_file_for_lipper_request.R
\name{create_isin_file_for_lipper_request}
\alias{create_isin_file_for_lipper_request}
\title{Create the csv file with the ISIN for the lipper SRRI request}
\usage{
create_isin_file_for_lipper_request(
  li_date,
  outDir = "//rackstation/MarcoShare/Per Panos/Lipper_richieste_srri"
)
}
\arguments{
\item{li_date}{the date of the li(yyyymmdd).RData file from which the li list
(list of investments) is constructed. This date determines the .RData file
used to filter the fund and similar investments whose ISIN is saved in the
output csv file to be imported in Lipper. li_date must be a date.}

\item{outDir}{the output directory where to save the data.}
}
\description{
Create the csv file with the ISIN for the lipper SRRI request
}
\examples{
\dontrun{
 li <- list()
create_isin_file_for_lipper_request(li, outDir = ".")
}
}
\author{
Claudio Ortelli
}
