% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRatingData.R
\name{createRatingData}
\alias{createRatingData}
\title{Create the rating definitions and rating maps used to identify the OC_SRRI proxy}
\usage{
createRatingData(asOfDate)
}
\arguments{
\item{asOfDate}{a valid date used to extract the rating map that was in use at asOfDate.}
}
\value{
a named list with the following variables:

- rating.l a named list with the rating definitions and rating classes for the 3 rating agencies.
- matClass.v a numeric named vector used to map the maturity of the fixed income instrument in the corresponding
maturity class
- allowedInvType.v a character vector of allowed investment types. Only the investment types in the vector are
allowed to be mapped by the 'by_rat_mat' proxy method (see the classifier definition in allocare AMS). Currently
the following inv. types are allowed: "Floater", "Perpetual Bonds", "Fix-To-Float", "Mid & Long Term Notes", "Ordinary Bonds", "High Yield Bonds", "Zero Bonds", "Other Bonds".
- mapByCcy.l a named list with the maps <rating, maturity> --> iid_proxy for the 3 currencies EUR, USD and CHF.
}
\description{
Create the rating definitions for rating agencies Moody's, Fitch an Standard and
Poors. The function returns the map from the pair <rating, maturity> to the iid_proxy
as well as the maturity classes used, i.e. [0,1), [1,3), ... .
}
\examples{
ratingData.l <- createRatingData(as.Date("2021-08-30"))
 
}
