% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeeklyReturnsTs.R
\name{createWeeklyReturnTs}
\alias{createWeeklyReturnTs}
\title{Compute the SRRI from a timeseries of returns and an asOfDate}
\usage{
createWeeklyReturnTs(i, usedProxy.tib, ts.l, li, lix)
}
\arguments{
\item{i}{an integer number between 1 and nrow(map.tib).}

\item{usedProxy.tib}{the tibble containing unique combinations of <iid_proxy, type_proxy>. Note that
the two columns of usedProxy.tib must have name 'iid' and 'type', respectively.}

\item{ts.l}{the result of spliting the list of default timeseries by factors 'Idtf', 'TSType' and 'PriceCcy'.}

\item{li}{the list of investments with the PXX tibbles.}

\item{lix}{the list of indices with the PXX tibbles.}
}
\value{
the xts of weekly returns. The attributes 'Ccy' and 'ts_type' are assigned to the returned xts.
}
\description{
Compute the SRRI from a timeseries of returns and an asOfDate
}
