% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{cycles}
\alias{cycles}
\title{Check if the map internal_id -> internal_id_proxy contains cycles}
\usage{
cycles(map.df, strict, verbose)
}
\arguments{
\item{map.df}{data.frame with character columns iid and iid_proxy.}

\item{strict}{logical. If TRUE a proxy can not have a proxy.}

\item{verbose}{logical. If true a sequence of info messages is printed.}
}
\value{
"cycles" is returned if one or more autoreferencing sets of 
        investments are identified. "no_cycles" otherwise.
}
\description{
Given a data.frame with columns iid and iid_proxy the implicit map  
iid -> iid_proxy is tested for cycles. The idea is to avoid 
autoreferencing  of investment objects when computing the OC_SRRI.
}
\examples{
h.df <- data.frame(
   iid = c("a", "b", "c", "e", "d", "v", "z", "k"),
   iid_proxy = c("b", "c", "k", "h", "p", "z", "s", "q"),
         stringsAsFactors = FALSE)
cycles(h.df, strict = FALSE, verbose = TRUE)
 
}
