% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAvailableInLipper.R
\name{extractAvailableInLipper}
\alias{extractAvailableInLipper}
\title{Extracts the ISIN code of the investments available in Lipper}
\usage{
extractAvailableInLipper(li, asOfDate, withInternalId = FALSE)
}
\arguments{
\item{li}{the list of investments information with P01 tibble.}

\item{asOfDate}{the asOfDate to be used to determine the list of available investment types.}

\item{withInternalId}{logical indicating if the INTERNAL_ID should be included in the output}
}
\value{
the character vector of ISIN codes of the available investments or a tibble with 
INTERNAL_ID and ISIN
}
\description{
Extracts the ISIN code of the investments available in Lipper
}
\examples{
li <- list(P01 = tibble::tibble(Description = "Stock Funds", ISIN = "LU0546265769"))
extractAvailableInLipper(li, as.Date("2021-09-20"))
 
}
