% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lipper_funds.R
\name{get_lipper_funds}
\alias{get_lipper_funds}
\title{Extract the INTERNAL_ID, ISIN pair of the investments objects available in Lipper}
\usage{
get_lipper_funds(asOfDate, li)
}
\arguments{
\item{asOfDate}{the as of date at which to extract the information}

\item{li}{list of investment objects from the Allocare request, i.e. a named list
with field names "P01" "P03" "P33" "P39" "P41" "P42" "P43" "P44" "P45" "P46" "P47" "P52".}
}
\value{
a tibble with columns INTERNAL_ID and ISIN.
}
\description{
Extract the INTERNAL_ID, ISIN pair of the investments objects available in Lipper
}
\examples{
\dontrun{
 li <- list()
get_lipper_funds(Sys.Date(), li)
}
}
\author{
Claudio Ortelli
}
