% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proxySRRI_OC.R
\name{get_proxySRRI_OC}
\alias{get_proxySRRI_OC}
\title{Get the SRRI_OC of the proxy assigned to an I.O.}
\usage{
get_proxySRRI_OC(id, asOfDate, li, lix)
}
\arguments{
\item{id}{investment object id (string).}

\item{asOfDate}{the as of date for the SRRI_OC.}

\item{li}{list of Pxx tibbles of investments data.}

\item{lix}{list of Pxx tibbles of index data.}
}
\value{
The SRRI_OC value (character) of the proxy linked to the I:O.
}
\description{
Get the SRRI_OC value of the proxy assigned to 
an investment object (I.O.) or NA if the I.O. 
does not have a proxy.
}
\examples{
 lix <- list('P41' = tibble::tibble('Idtf' = '38668', 'RatingAgency' = 'SRRI_OC',
   'RatingCode' = '5', 'StartDate' = as.Date('2021-01-09'), 'Expiry' = as.Date(NA_character_)))

 li <- list('P42' = tibble::tibble('Idtf' = '12345', 'IndClass' = 'OC_SRRI_proxy',
   'IndCode' = '38668', 'StartDate' = as.Date('2021-01-09'), 'Expiry' = as.Date(NA_character_)),
     'P41' = tibble::tibble('Idtf' = '552234', 'RatingAgency' = 'SRRI_OC',
        'RatingCode' = '7', 'StartDate' = as.Date('2021-01-09'), 'Expiry' = as.Date(NA_character_)))


 get_proxySRRI_OC('12345', as.Date('2021-01-09'), li, lix)
}
