% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProxyMap.R
\name{getProxyMap}
\alias{getProxyMap}
\title{Get the map between iid and iid_proxy}
\usage{
getProxyMap(li)
}
\arguments{
\item{li}{the 'list of investments' containing the P42 tibble.}
}
\value{
a tibble with 'iid', 'type', 'iid_proxy', 'type_proxy', 'StartDate', 'Expiry' columns
where iid stands for INTERNAL_ID and 'type' denotes the type of the investment, i.e. `inv` (investment) or
`idx` (index).
}
\description{
This function constructs the map between the investment objects and the corresponding 
proxy used in SRRI calculation. The 'li' list of instruments argument contains all information available at a given point in time. Usually
the 'li' is the current (today) list of investment objects.
}
\details{
Get the map between iid and iid_proxy
}
