% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_li_rdata.R
\name{load_li_rdata}
\alias{load_li_rdata}
\title{Load list of investments in a li(yyyymmdd).RData file}
\usage{
load_li_rdata(li_date)
}
\arguments{
\item{li_date}{the date of the li(yyyymmdd).RData file from which the li list 
(list of investments) is constructed. This date determines the .RData file
used to filter the fund and similar investments whose ISIN is saved in the
output csv file to be imported in Lipper. li_date must be a date.}
}
\value{
li list of investment objects from the Allocare request, i.e. a named list
        with field names "P01" "P03" "P33" "P39" "P41" "P42" "P43" "P44" "P45" "P46" "P47" "P52".
}
\description{
Load list of investments in a li(yyyymmdd).RData file
}
\examples{
\dontrun{
li <- load_li_rdata(Sys.Date())
}
}
\author{
Claudio Ortelli
}
