% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noTsNeeded.R
\name{noTsNeeded}
\alias{noTsNeeded}
\title{Identifies the special proxies that do not require a timeseries for srri calculation}
\usage{
noTsNeeded(map.tib)
}
\arguments{
\item{map.tib}{a tibble with a 'iid_proxy' field.}
}
\value{
a logical vector of the same length as map.tib[['iid_proxy']]
}
\description{
Identifies the special proxies that do not require a timeseries for srri calculation
}
