% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyConsistencyCheck.R
\name{proxyConsistencyCheck}
\alias{proxyConsistencyCheck}
\title{Check consistency of a given map.tib at a reference date}
\usage{
proxyConsistencyCheck(asOfDate, map.tib, li)
}
\arguments{
\item{asOfDate}{the reference date used to adjust the map.tib.}

\item{map.tib}{the map between investment iid and proxy iid. the map.tib is
defined by means of the following 6 columns 'iid', 'type', 'iid_proxy', 'type_proxy',
'StartDate', 'Expiry'}

\item{li}{the exported list of investments containing the PXX tibbles.}
}
\value{
an empty data.frame(iid = character(0), iid_proxy = character(0), stringsAsFactors = FALSE) if
no cycles are detected, the iid -> iid_proxy data.frame map of the involved investments otherwise.
}
\description{
Check consistency of a given map.tib at a reference date
}
