% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyMapByTimeRange.R
\name{proxyMapByTimeRange}
\alias{proxyMapByTimeRange}
\title{Get the internal identifiers of the used proxy SRRI_OC}
\usage{
proxyMapByTimeRange(dt, ndays = 7L * 260L, map.tib)
}
\arguments{
\item{dt}{the as of date for the request.}

\item{ndays}{the number of days in the past that must be considered in the identification
of the used proxy. Use ndays = 0 for an atomic selection based on the asOfDate dt only.}

\item{map.tib}{the map between investments and proxies. The following columns are needed: 
'iid', 'type', 'iid_proxy', 'type_proxy', 'StartDate', 'Expiry'}
}
\value{
A four columns tibble with 'iid', 'type', 'iid_proxy' and 'type_proxy' 
defining the map between I.O./Index and corresponding proxy. 
No particular check is performed to check for circularity in the proxy assignment between 
investment objects. If you need to avoid circular autoreferencing use the 
proxyValidityCheck function.
}
\description{
Given an asOfDate and the required connection parameters the internal identifiers of the 
used proxy SRRI_OC are returned
}
\examples{
\dontrun{
proxyMapByTimeRange(dt = Sys.Date(), ndays = 7L*260L, map.tib)
} 
}
