% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyMapValidityCheck.R
\name{proxyMapValidityCheck}
\alias{proxyMapValidityCheck}
\title{Check the validity of SRRI_OC_proxy identifiers}
\usage{
proxyMapValidityCheck(proxyMap.tib, strict = TRUE, action)
}
\arguments{
\item{proxyMap.tib}{a tibble with 2 columns <Int_id, Proxy_Int_id>}

\item{strict}{logical. If TRUE proxy may not have a proxy.}

\item{action}{character string equal to "warn" or "stop" determining
the behaviour of the function in case of invalid entries (i.e. I.O. used
as proxy having themself a proxy). If action = 'stop' the function stops
with an error message. If action = 'warn' a warning is produced but the
function continues the execution. If action is missing no warning and no
error is reportded.}
}
\value{
A named list with
   - info a character string with possible values "no_cycles" or "cycles"
   - map.df the data.frame with the mapping iid -> iid_proxy of the self
     referencing investments.
}
\description{
Check the validity of SRRI_OC_proxy identifiers by requiring that
an investment object or an index used as proxy by another I.O. or
index can not have a proxy.
}
\examples{
proxyMap.tib <- data.frame(iid = c("1", "2", "3"), type = rep("inv", 3), 
                    iid_proxy = c("2", "3", "5"), type_proxy = rep("idx", 3), 
                    stringsAsFactors = FALSE)
result.l <- proxyMapValidityCheck(proxyMap.tib, strict = TRUE, action = 'warn')
}
