% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_srri.R
\name{push_srri}
\alias{push_srri}
\title{Compute and push the SRRI to Allocare AMS}
\usage{
push_srri(
  li_date,
  lipper_xlsx_date,
  lipper_srri_dir = "//rackstation/MarcoShare/Per Panos/Lipper_richieste_srri",
  dryrun = FALSE
)
}
\arguments{
\item{li_date}{the date of the li(yyyymmdd).RData file from which the li list
(list of investments) is constructed. This date determines the .RData file
used to filter the fund and similar investments whose ISIN is saved in the
output csv file to be imported in Lipper. li_date must be a date.}

\item{lipper_xlsx_date}{the date identifying the xlsx file to load. By convention
the file name must be equal to 'lista lipper srri fondi yyyy-mm-dd.xlsx'.}

\item{lipper_srri_dir}{the directory where to search for the lipper xlsx file-}

\item{dryrun}{if TRUE the data are not push to Allocare.}
}
\value{
a named list with the following slots:
    p41.tib (the data pushed to AMS), computed.tib, fund.tib, cash.tib, ccyFwd.tib, unassigned.tib
}
\description{
Compute and push the SRRI to Allocare AMS
}
\examples{
\dontrun{
result_l <- push_srri(Sys.Date(), Sys.Date() - 2, lipper_srri_dir = ".", dryrun = TRUE)
}
}
\author{
Claudio Ortelli
}
