% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_srri_with_proxy.R
\name{update_srri_with_proxy}
\alias{update_srri_with_proxy}
\title{Update the SRRI of investiment object with the SRRI of the proxy if defined.}
\usage{
update_srri_with_proxy(li_date, push = FALSE, outDir = ".", verbose = FALSE)
}
\arguments{
\item{li_date}{the date of the li(yyyymmdd).RData file from which the li list
(list of investments) is constructed. This date determines the .RData file
used to filter the fund and similar investments whose ISIN is saved in the
output csv file to be imported in Lipper. li_date must be a date.}

\item{push}{should the result be pushed to Allocare. Default to FALSE. Functionality not
working yet. Function stops with error message if push = TRUE.}

\item{outDir}{The output directory where to write the 'invalidMapElements.RData' file.
Defaults to ".".}

\item{verbose}{Should the INFO messages be displayed? Defaults to FALSE.}
}
\value{
a named list with the following tibbles: tsNeeded.tib, staticSrri.tib,
                noChange.tib, lipper.tib.
}
\description{
Update the SRRI of investiment object with the SRRI of the proxy if defined.
}
\examples{
\dontrun{
result_l <- update_srri_with_proxy(li_date = Sys.Date() - 1, push = FALSE,
             outDir = ".", verbose = TRUE)
}
}
\author{
Claudio Ortelli
}
