% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volToSrri.R
\name{volToSrri}
\alias{volToSrri}
\title{Map yearly realized volatility to SRRI}
\usage{
volToSrri(yearlyVol)
}
\arguments{
\item{yearlyVol}{the yearly realized volatility computed with weekly returns over a 5 years period (260 obs.).}
}
\value{
the SRRI figure. See CERs Guidelines 10-673 on the methodology for the calculation of the synthetic risk and reward indicator.
}
\description{
Map yearly realized volatility to SRRI
}
\examples{
srri <- volToSrri(0.15)
 
}
